Move home page content to "home.svelte"
This frees up "index.svelte" to use for initial user auth on arrival.
The contents of "home.svelte" and "index.svelte" were merely swapped
unchanged, to keep this PR simple to review.
